Fixes error where an attempt is made to delete a non-existent old schedule.

+++ com_organizer/site/Models/ScheduleJSON.php
@@ -757,7 +757,10 @@ class ScheduleJSON extends BaseModel
 		$this->setUnitReferences();
 		$this->setCalendarReferences();
 
-		$reference->delete();
+		if ($reference->id)
+		{
+			$reference->delete();
+		}
 	}
 
 	/**
